reduce the process name pipe
