--- lib/process-services-cloud/src/lib/pipes/process-services-cloud-pipe.module.ts
@@ -1,30 +1,0 @@
-/*!
- * @license
- * Copyright © 2005-2024 Hyland Software, Inc. and its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { NgModule } from '@angular/core';
-import { ProcessNameCloudPipe } from './process-name-cloud.pipe';
-
-@NgModule({
-    declarations: [
-        ProcessNameCloudPipe
-    ],
-    exports: [
-        ProcessNameCloudPipe
-    ]
-})
-export class ProcessServicesCloudPipeModule {
-}
