+++ lib/process-services-cloud/src/lib/process/start-process/components/start-process-cloud.component.spec.ts
@@ -46,7 +46,6 @@ import { By } from '@angular/platform-browser';
 import { ProcessPayloadCloud } from '../models/process-payload-cloud.model';
 import { ProcessWithFormPayloadCloud } from '../models/process-with-form-payload-cloud.model';
 import { ProcessServiceCloudTestingModule } from '../../../testing/process-service-cloud.testing.module';
-import { ProcessNameCloudPipe } from '../../../pipes/process-name-cloud.pipe';
 import { ProcessInstanceCloud } from '../models/process-instance-cloud.model';
 import { ESCAPE } from '@angular/cdk/keycodes';
 import { ProcessDefinitionCloud } from '../../../models/process-definition-cloud.model';
@@ -999,8 +998,7 @@ describe('StartProcessCloudComponent', () => {
         });
 
         it('should set the process name using the processName cloud pipe when a process definition gets selected', async () => {
-            const processNameCloudPipe = TestBed.inject(ProcessNameCloudPipe);
-            const processNamePipeTransformSpy = spyOn(processNameCloudPipe, 'transform').and.returnValue('fake-transformed-name');
+            const getDefaultProcessNameSpy = spyOn(component, 'getDefaultProcessName').and.returnValue('fake-transformed-name');
             const expectedProcessInstanceDetails: ProcessInstanceCloud = { processDefinitionName: fakeProcessDefinitions[0].name };
             getDefinitionsSpy.and.returnValue(of([fakeProcessDefinitions[0]]));
             formDefinitionSpy.and.stub();
@@ -1011,7 +1009,7 @@ describe('StartProcessCloudComponent', () => {
             fixture.detectChanges();
             await fixture.whenStable();
 
-            expect(processNamePipeTransformSpy).toHaveBeenCalledWith(component.name, expectedProcessInstanceDetails);
+            expect(getDefaultProcessNameSpy).toHaveBeenCalledWith(component.name, expectedProcessInstanceDetails);
             expect(component.processInstanceName.dirty).toBe(true);
             expect(component.processInstanceName.touched).toBe(true);
             expect(component.processInstanceName.value).toEqual('fake-transformed-name');
@@ -1038,9 +1036,8 @@ describe('StartProcessCloudComponent', () => {
 
             getDefinitionsSpy.and.returnValue(of(definitions));
 
-            const processNameCloudPipe = TestBed.inject(ProcessNameCloudPipe);
             const fakeTransformedName = 'fake-transformed-name';
-            spyOn(processNameCloudPipe, 'transform').and.returnValue(fakeTransformedName);
+            spyOn(component, 'getDefaultProcessName').and.returnValue(fakeTransformedName);
 
             component.processDefinitionName = 'fake-name';
             const change = new SimpleChange(null, 'MyApp', true);
