+++ lib/process-services-cloud/src/lib/process/start-process/components/start-process-cloud.component.ts
@@ -29,14 +29,7 @@ import {
     ViewChild,
     ViewEncapsulation
 } from '@angular/core';
-
-import {
-    ContentLinkModel,
-    FORM_FIELD_VALIDATORS,
-    FormFieldValidator,
-    FormModel,
-    TranslationService
-} from '@alfresco/adf-core';
+import { ContentLinkModel, FORM_FIELD_VALIDATORS, FormFieldValidator, FormModel, LocalizedDatePipe, TranslationService } from '@alfresco/adf-core';
 import { AbstractControl, FormControl, FormGroup, ValidatorFn, Validators } from '@angular/forms';
 import { MatAutocompleteTrigger } from '@angular/material/autocomplete';
 import { catchError, debounceTime } from 'rxjs/operators';
@@ -47,12 +40,14 @@ import { StartProcessCloudService } from '../services/start-process-cloud.servic
 import { forkJoin, of } from 'rxjs';
 import { ProcessDefinitionCloud } from '../../../models/process-definition-cloud.model';
 import { TaskVariableCloud } from '../../../form/models/task-variable-cloud.model';
-import { ProcessNameCloudPipe } from '../../../pipes/process-name-cloud.pipe';
 import { FormCloudDisplayModeConfiguration } from '../../../services/form-fields.interfaces';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { getTime } from 'date-fns';
 
 const MAX_NAME_LENGTH: number = 255;
 const PROCESS_DEFINITION_DEBOUNCE: number = 300;
+const DATE_TIME_IDENTIFIER_REG_EXP = new RegExp('%{datetime}', 'i');
+const PROCESS_DEFINITION_IDENTIFIER_REG_EXP = new RegExp('%{processdefinition}', 'i');
 
 @Component({
     selector: 'adf-cloud-start-process',
@@ -140,7 +135,6 @@ export class StartProcessCloudComponent implements OnChanges, OnInit {
     resolvedValues?: TaskVariableCloud[];
     customOutcome: string;
 
-
     isProcessStarting = false;
     isFormCloudLoaded = false;
     isFormCloudLoading = false;
@@ -162,7 +156,7 @@ export class StartProcessCloudComponent implements OnChanges, OnInit {
 
     private readonly destroyRef = inject(DestroyRef);
     private readonly startProcessCloudService = inject(StartProcessCloudService);
-    private readonly processNameCloudPipe = inject(ProcessNameCloudPipe);
+    private readonly localizedDatePipe = inject(LocalizedDatePipe);
 
     get isProcessFormValid(): boolean {
         if (this.hasForm && this.isFormCloudLoaded) {
@@ -500,9 +494,23 @@ export class StartProcessCloudComponent implements OnChanges, OnInit {
 
     setDefaultProcessName(processDefinitionName: string): void {
         const processInstanceDetails: ProcessInstanceCloud = { processDefinitionName };
-        const defaultProcessName = this.processNameCloudPipe.transform(this.name, processInstanceDetails);
+        const defaultProcessName = this.getDefaultProcessName(this.name, processInstanceDetails);
         this.processInstanceName.setValue(defaultProcessName);
         this.processInstanceName.markAsDirty();
         this.processInstanceName.markAsTouched();
+    }
+
+    getDefaultProcessName(processNameFormat: string, processInstance?: ProcessInstanceCloud): string {
+        let processName = processNameFormat;
+        if (processName.match(DATE_TIME_IDENTIFIER_REG_EXP)) {
+            const presentDateTime = getTime(new Date());
+            processName = processName.replace(DATE_TIME_IDENTIFIER_REG_EXP, this.localizedDatePipe.transform(presentDateTime, 'medium'));
+        }
+
+        if (processName.match(PROCESS_DEFINITION_IDENTIFIER_REG_EXP)) {
+            const selectedProcessDefinitionName = processInstance ? processInstance.processDefinitionName : '';
+            processName = processName.replace(PROCESS_DEFINITION_IDENTIFIER_REG_EXP, selectedProcessDefinitionName);
+        }
+        return processName;
     }
 }
