+++ lib/process-services-cloud/src/public-api.ts
@@ -27,8 +27,6 @@ export * from './lib/rich-text-editor/public-api';
 
 export * from './lib/types';
 export * from './lib/common/index';
-export * from './lib/pipes/process-name-cloud.pipe';
-export * from './lib/pipes/process-services-cloud-pipe.module';
 export * from './lib/models/process-definition-cloud.model';
 export * from './lib/models/date-cloud-filter.model';
 export * from './lib/models/application-version.model';
@@ -39,4 +37,4 @@ export * from './lib/models/task-list-sorting.model';
 export * from './lib/models/process-instance-variable.model';
 export * from './lib/models/variable-definition';
 export * from './lib/models/date-format-cloud.model';
-export * from './lib/models/process-variable-filter.model'
+export * from './lib/models/process-variable-filter.model';
